Add an initial README and piazza-api dependency.
piazza-api is in a vendor directory specifically because the project
does not appear to have a Python 3 wheel available on PyPI for v0.7.0.
Nothing about the project seems to be Python 2 specific, so for now the
easiest thing to do is to just bundle it as a git submodule. Further
dependencies will be managed via pipenv.