Add initial demo for crawling progress bar.
This involved a lot of effort in re-architecting the app to use redis as
the message queue backend for both celery and the SocketIO server.
Things get a little hairy with eventlet monkey-patching, so the app now
uses a factory function to make SocketIO use the eventlet or threading
backend, depending on its parameter. For the regular webserver + socket
server, eventlet is used and the proper entry point is wsgi.py. For the
celery worker, a new entry point at celery.py was added that does _not_
monkey-patch eventlet, as otherwise a deadlock occurs.

RabbitMQ was not used as the backend because the RPC result backend
could not properly give status updates to the socket server or the web
server outside of the initial request that created the task. Since we
need the status of the task for progress reporting, this wasn't
acceptable to me, so we switched to redis entirely.